BAU: Stop using sed, use %x over backticks
- Use a hash instead of nasty sed
- Use %x and capture output/result of call for error checking
- Provide nicer output of what certs are being generated

+++ generate/pki.rb
@@ -1,6 +1,7 @@
 #!/usr/bin/env ruby
 
 require 'yaml'
+require 'json'
 require 'fileutils'
 
 SCRIPT_DIR = File.expand_path(File.dirname(__FILE__))
@@ -9,31 +10,51 @@ pki_dir = 'pki'
 FileUtils::mkdir_p pki_dir
 Dir.chdir pki_dir
 
+def get_template_csr(common_name)
+  {
+      "CN": common_name,
+      "hosts": [ "localhost" ],
+      "key": {
+          "algo": "rsa",
+          "size": 2048
+      },
+      "names": [
+          {
+              "C": "GB",
+              "ST": "London",
+              "L": "London",
+              "O": "Cabinet Office",
+              "OU": "GDS"
+          }
+      ]
+  }.to_json
+end
+
 def generate_cert(type, common_name, file, ca_cert, ca_key)
-    command = <<-COMMAND
-      sed 's/$COMMON_NAME/#{common_name}/' #{SCRIPT_DIR}/template-csr.json | \
-      cfssl gencert -config #{SCRIPT_DIR}/cfssl-config.json -profile "#{type}" -ca #{ca_cert} -ca-key #{ca_key} /dev/stdin | \
-      cfssljson -bare #{file}
-    COMMAND
-    `#{command}`
+  output = %x(echo '#{get_template_csr(common_name)}' | \
+              cfssl gencert -config #{SCRIPT_DIR}/cfssl-config.json -profile "#{type}" -ca #{ca_cert} -ca-key #{ca_key} -loglevel=2 - | \
+              cfssljson -bare #{file})
+  raise output unless $?.success?
 end
 
 def convert_key_cert(file)
-    `openssl pkcs8 -topk8 -inform PEM -outform DER -in #{file}-key.pem -out #{file}.pk8 -nocrypt`
-    FileUtils.move "#{file}.pem", "#{file}.crt"
-    FileUtils.remove [ "#{file}-key.pem", "#{file}.csr" ]
+  openssl_output = %x(openssl pkcs8 -topk8 -inform PEM -outform DER -in #{file}-key.pem -out #{file}.pk8 -nocrypt)
+  raise openssl_output unless $?.success?
+  FileUtils.move "#{file}.pem", "#{file}.crt"
+  FileUtils.remove [ "#{file}-key.pem", "#{file}.csr" ]
 end
 
 pki_to_generate['pki'].each do |ca_name, ca|
-  puts "Generate key/cert for Root CA: #{ca_name}"
-  `sed 's/$COMMON_NAME/#{ca['common-name']}/' #{SCRIPT_DIR}/template-csr.json | cfssl genkey -initca /dev/stdin | cfssljson -bare #{ca['cert-file']}`
+  puts "Generating Root CA: #{ca_name}"
+  output = %x(echo '#{get_template_csr(ca['common-name'])}' | cfssl genkey -initca -loglevel=2 - | cfssljson -bare #{ca['cert-file']})
+  raise output unless $?.success?
 
   ca['intermediates'].each do |inter_name, inter|
     # Need to change values in the hash so that future access also gets these defaults
     inter['common-name'] = inter.fetch('common-name', "Verify Local Startup #{inter_name.capitalize} Intermediate CA")
     inter['cert-file'] = inter.fetch('cert-file', "#{ca_name}-#{inter_name}")
 
-    puts "Generate key/cert for intermediate CA: #{inter_name}"
+    puts "\tGenerating Intermediate CA: #{inter_name}"
     generate_cert('intermediate',
                   inter['common-name'],
                   inter['cert-file'],
@@ -42,7 +63,8 @@ pki_to_generate['pki'].each do |ca_name, ca|
 
     inter['certs'].each do |cert_name, cert|
       cert_file = cert.fetch('cert-file', cert_name)
-      common_name = cert.fetch('common-name', "Verify Local Startup #{cert_name.split("-").map{ |s| s.capitalize}.join(' ')}"),
+      common_name = cert.fetch('common-name', "Verify Local Startup #{cert_name.split("-").map{ |s| s.capitalize}.join(' ')}")
+      puts "\t\tGenerating cert: #{cert_name}"
       generate_cert(cert['type'],
                     common_name,
                     cert_file,
